Allow escaping characters using backslashes
A backslash can be inserted in order to escape a character in the macro
parameters and give it its literal meaning. Typically this would be a comma (to
include the comma in the substitution rather than using it to separate multiple
parameters). Backslashes themselves can be escaped with more backslashes.